+++ src/main/java/lotto/domain/LottoTicket.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import lotto.domain.validator.Validator;
+import lotto.validator.Validator;
 
 import java.util.Collections;
 import java.util.List;
